requirements.txt
fastapi
uvicorn
SQLAlchemy
asyncpg
alembic
pydantic
python-dotenv
